requirements.txt
click==8.1.7
python-dotenv==1.0.1
requests==2.31.0